requirements.txt
pycrypto
wave